patient add/delete/edit

+++ backend/app/api/medicine/delete/route.js
@@ -6,6 +6,13 @@ export async function POST(req) {
         const body = await req.json();
         const {id} = body;
 
+        if (!id || id === "") {
+            return NextResponse.json(
+                {error: "Drug id is empty", status: 204},
+                {status: 204}
+            )
+        }
+
         const status = await sql`delete from drugs where drug_id = ${id}`;
         return NextResponse.json(status);
     }
